Allow to sort stake delegates by alignment or total delegated
